feat(seo): add schema-org
