+++ src/components/SEO/seo.tsx
@@ -3,14 +3,16 @@ import { graphql, useStaticQuery } from 'gatsby';
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { OpenGraph } from './open-graph';
+import { SchemaOrg } from './schema-org';
 import { Twitter } from './twitter';
 
 type Props = {
   title?: Maybe<string>;
   image?: Maybe<string>;
   description?: Maybe<string>;
+  isBlogPost?: boolean;
   blogSlug?: Maybe<string>;
-  article?: boolean;
+  datePublished?: Maybe<string>;
 };
 
 type PureProps = Props & { site: Site };
@@ -19,9 +21,10 @@ export const PureSEO: React.FC<PureProps> = ({
   title,
   description,
   image,
-  article,
+  isBlogPost = false,
   site,
   blogSlug,
+  datePublished,
 }) => {
   const {
     defaultTitle,
@@ -31,17 +34,19 @@ export const PureSEO: React.FC<PureProps> = ({
     siteUrl,
     twitter,
     lang,
+    author,
+    organization,
   } = site.siteMetadata;
 
-  const metaImage = `${siteUrl}${image || defaultImage}`;
-  const metaDescription = description || defaultDescription;
+  const metaImage = `${siteUrl}${image ?? defaultImage}`;
+  const metaDescription = description ?? defaultDescription;
   const url = blogSlug ? `${siteUrl}${blogSlug}` : siteUrl;
 
   return (
     <>
       <Helmet
-        defaultTitle={defaultTitle || ''}
-        titleTemplate={titleTemplate || ''}
+        defaultTitle={defaultTitle ?? ''}
+        titleTemplate={titleTemplate ?? ''}
       >
         {title && <title>{title}</title>}
         {lang && <html lang={lang} />}
@@ -55,13 +60,25 @@ export const PureSEO: React.FC<PureProps> = ({
         title={title}
         description={metaDescription}
         image={metaImage}
-        article={article}
+        isBlogPost={isBlogPost}
       />
       <Twitter
         twitter={twitter}
         title={title}
         description={metaDescription}
         image={metaImage}
+      />
+      <SchemaOrg
+        url={url}
+        title={title ?? defaultTitle}
+        defaultTitle={defaultTitle}
+        isBlogPost={isBlogPost}
+        image={metaImage}
+        canonicalUrl={siteUrl}
+        author={author}
+        description={description}
+        organization={organization}
+        datePublished={datePublished}
       />
     </>
   );
@@ -79,6 +96,12 @@ export const SEO: React.FC<Props> = props => {
           lang
           siteUrl
           twitter
+          author
+          organization {
+            name
+            logo
+            url
+          }
         }
       }
     }
@@ -87,8 +110,4 @@ export const SEO: React.FC<Props> = props => {
   const data = useStaticQuery<{ site: Site }>(query);
 
   return <PureSEO {...data} {...props} />;
-};
-
-SEO.defaultProps = {
-  article: false,
 };
